Moved InfobloxNetmri class outside of __init__.py
This makes the package layout more pythonic (e.g. only imports and __all__ in __init__.py)
